# forum-mse-2026/src/main/resources/openapi/openapi.yaml
openapi: 3.0.3

info:
  title: Forum API
  description: REST API for forum posts and users (JWT bearer auth).
  version: 1.0.1

servers:
  - url: http://localhost:9000
    description: Local

tags:
  - name: Auth
    description: Authentication
  - name: Posts
    description: Post operations
  - name: Replies
    description: Replies on posts
  - name: Users
    description: User operations

paths:
  /auth/register:
    post:
      tags: [Auth]
      operationId: register
      summary: Register a regular form user
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RegisterResponse'
        '400':
          description: Invalid request
        '409':
          description: Username already taken

  /auth/login:
    post:
      tags: [Auth]
      operationId: login
      summary: Obtain a JWT (username + password)
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResponse"
        "401":
          description: Invalid credentials

  /posts:
    get:
      tags: [Posts]
      operationId: listPosts
      summary: List all posts
      security: []
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PostResponse"

    post:
      tags: [Posts]
      operationId: createPost
      summary: Create a post
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePostRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PostResponse"
        '400':
          description: Invalid request
        '401':
          description: Authentication required
        '409':
          description: Title alreadyt exists

  /posts/{id}:
    get:
      tags: [Posts]
      operationId: getPostById
      summary: Get post by id
      security: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: Found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PostResponse"
        "404":
          description: Not found
    put:
      tags: [Posts]
      operationId: updatePost
      summary: Update a post (topic)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdatePostRequest'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PostResponse'
        '400':
          description: Invalid request
        '403':
          description: Forbidden
        '404':
          description: Not FOund
        '409':
          description: Title already exists

  /posts/{postId}/replies:
    get:
      tags: [Replies]
      operationId: listRepliesForPost
      summary: List replies for a post
      security: []
      parameters:
        - name: postId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ReplyResponse"
        "404":
          description: Post not found

    post:
      tags: [Replies]
      operationId: createReply
      summary: Create a reply on a post
      parameters:
        - name: postId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateReplyRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReplyResponse"
        "404":
          description: Post not found

  /replies/{id}:
    get:
      tags: [Replies]
      operationId: getReplyById
      summary: Get reply by id
      security: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: Found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReplyResponse"
        "404":
          description: Not found

  /users:
    get:
      tags: [Users]
      operationId: listUsers
      summary: List all users
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/UserResponse"

    post:
      tags: [Users]
      operationId: createUser
      summary: Create a user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateUserRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserResponse"
        "409":
          description: Username or email already in use

  /users/{id}:
    get:
      tags: [Users]
      operationId: getUserById
      summary: Get user by id
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: Found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserResponse"
        "404":
          description: Not found

    put:
      tags: [Users]
      operationId: updateUser
      summary: Update a user
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUserRequest"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserResponse"
        "404":
          description: Not found
        "409":
          description: Username or email already in use

    delete:
      tags: [Users]
      operationId: deleteUser
      summary: Delete a user
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "204":
          description: No content
        "404":
          description: Not found

  /users/{id}/role:
    patch:
      tags: [Users]
      operationId: updateUserRole
      summary: Update user role req.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateUserRoleRequest'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '403':
          description: Forbidden
        '404':
          description: Not Founds


components:

  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  schemas:
    RegisterRequest:
      type: object
      required: [username, password]
      properties:
        username:
          type: string
          minLength: 1
          maxLength: 100
        password:
          type: string
          minLength: 8
          maxLength: 128

    RegisterResponse:
      type: object
      required: [id, username]
      properties:
        id:
          type: integer
          format: int64
        username:
          type: string

    LoginRequest:
      type: object
      required: [username, password]
      properties:
        username:
          type: string
          minLength: 1
          maxLength: 100
        password:
          type: string
          minLength: 1
          maxLength: 128

    LoginResponse:
      type: object
      required: [accessToken, tokenType, expiresInSeconds]
      properties:
        accessToken:
          type: string
        tokenType:
          type: string
          example: Bearer
        expiresInSeconds:
          type: integer
          format: int64

    UserRole:
      type: string
      description: USER denotes a regular (non-privileged) account.
      enum: [ADMIN, MODERATOR, USER]

    UpdateUserRoleRequest:
      type: object
      required: [role]
      properties:
        role:
          $ref: '#/components/schemas/UserRole'

    CreateUserRequest:
      type: object
      required: [username, role, password]
      properties:
        username:
          type: string
          minLength: 1
          maxLength: 100
        email:
          type: string
          format: email
          maxLength: 320
        role:
          $ref: "#/components/schemas/UserRole"
        password:
          type: string
          minLength: 8
          maxLength: 72
          description: Stored as a bcrypt hash; max 72 bytes for bcrypt.

    UpdateUserRequest:
      type: object
      required: [username, role]
      properties:
        username:
          type: string
          minLength: 1
          maxLength: 100
        email:
          type: string
          format: email
          maxLength: 320
          nullable: true
        role:
          $ref: "#/components/schemas/UserRole"
        password:
          type: string
          minLength: 8
          maxLength: 72
          description: If set, replaces the password.

    UserResponse:
      type: object
      required: [id, username, role, createdAt]
      properties:
        id:
          type: integer
          format: int64
        username:
          type: string
        email:
          type: string
          format: email
          nullable: true
        role:
          $ref: "#/components/schemas/UserRole"
        createdAt:
          type: string
          format: date-time

    CreatePostRequest:
      type: object
      required: [title, content]
      properties:
        title:
          type: string
          maxLength: 500
        content:
          type: string
          maxLength: 10000

    PostResponse:
      type: object
      required: [ id, title, content, createdByUserId, createdByUsername, createdAt, updatedAt, viewCount ]
      properties:
        id:
          type: integer
          format: int64
        title:
          type: string
        content:
          type: string
        createdByUserId:
          type: integer
          format: int64
        createdByUsername:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        viewCount:
          type: integer
          format: int64

    UpdatePostRequest:
      type: object
      required: [title, content]
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 500
        content:
          type: string
          minLength: 1
          maxLength: 1000 # capped at 1000


    CreateReplyRequest:
      type: object
      required: [content]
      properties:
        content:
          type: string
          maxLength: 10000

    ReplyResponse:
      type: object
      required: [ id, postId, text, createdByUserId, createdByUsername, createdAt, updatedAt ]
      properties:
        id:
          type: integer
          format: int64
        postId:
          type: integer
          format: int64
        text:
          type: string
        createdByUserId:
          type: integer
          format: int64
        createdByUsername:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time

    PageReplyResponse:
      type: object
      required: [ items, page, size, totalElements, totalPages ]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/ReplyResponse'
        page:
          type: integer
          format: int32
        size:
          type: integer
          format: int32
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
          format: int32

security:
  - bearerAuth: []